requirements.txt
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
configparser==7.0.0
cryptography==42.0.8
Flask==3.0.3
Flask-Cors==4.0.1
google-api-core==2.19.1
google-auth==2.30.0
google-cloud-vision==3.7.2
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.62.2
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jwt==1.3.1
MarkupSafe==2.1.5
packaging==24.1
pillow==10.3.0
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
requests==2.32.3
rsa==4.9
urllib3==2.2.2
Werkzeug==3.0.3
